support detection bbox

+++ 256.192.model/test.py
@@ -51,12 +51,12 @@ def main(args):
         with torch.no_grad():
             input_var = torch.autograd.Variable(inputs.cuda())
             if args.flip == True:
-    	        flip_inputs = inputs.clone()
-    	        for i, finp in enumerate(flip_inputs):
-    	            finp = im_to_numpy(finp)
-    	            finp = cv2.flip(finp, 1)
-    	            flip_inputs[i] = im_to_torch(finp)
-            	flip_input_var = torch.autograd.Variable(flip_inputs.cuda())
+                flip_inputs = inputs.clone()
+                for i, finp in enumerate(flip_inputs):
+                    finp = im_to_numpy(finp)
+                    finp = cv2.flip(finp, 1)
+                    flip_inputs[i] = im_to_torch(finp)
+                flip_input_var = torch.autograd.Variable(flip_inputs.cuda())
 
             # compute output
             global_outputs, refine_output = model(input_var)
@@ -64,19 +64,19 @@ def main(args):
             score_map = score_map.numpy()
 
             if args.flip == True:
-    	        flip_global_outputs, flip_output = model(flip_input_var)
-    	        flip_score_map = flip_output.data.cpu()
-    	        flip_score_map = flip_score_map.numpy()
+                flip_global_outputs, flip_output = model(flip_input_var)
+                flip_score_map = flip_output.data.cpu()
+                flip_score_map = flip_score_map.numpy()
 
-    	        for i, fscore in enumerate(flip_score_map):
-    	            fscore = fscore.transpose((1,2,0))
-    	            fscore = cv2.flip(fscore, 1)
-    	            fscore = list(fscore.transpose((2,0,1)))
-    	            for (q, w) in cfg.symmetry:
-    	               fscore[q], fscore[w] = fscore[w], fscore[q] 
-    	            fscore = np.array(fscore)
-    	            score_map[i] += fscore
-    	            score_map[i] /= 2
+                for i, fscore in enumerate(flip_score_map):
+                    fscore = fscore.transpose((1,2,0))
+                    fscore = cv2.flip(fscore, 1)
+                    fscore = list(fscore.transpose((2,0,1)))
+                    for (q, w) in cfg.symmetry:
+                       fscore[q], fscore[w] = fscore[w], fscore[q] 
+                    fscore = np.array(fscore)
+                    score_map[i] += fscore
+                    score_map[i] /= 2
 
             ids = meta['imgID'].numpy()
             det_scores = meta['det_scores']
@@ -122,7 +122,7 @@ def main(args):
                     single_result_dict['image_id'] = int(ids[b])
                     single_result_dict['category_id'] = 1
                     single_result_dict['keypoints'] = single_result
-                    single_result_dict['score'] = det_scores[b]*v_score.mean()
+                    single_result_dict['score'] = double(det_scores[b])*v_score.mean()
                     full_result.append(single_result_dict)
 
     result_path = args.result
@@ -147,11 +147,11 @@ if __name__ == '__main__':
     parser.add_argument('-c', '--checkpoint', default='checkpoint', type=str, metavar='PATH',
                         help='path to load checkpoint (default: checkpoint)')
     parser.add_argument('-f', '--flip', default=True, type=bool,
-    					help='flip input image during test (default: True)')
+                        help='flip input image during test (default: True)')
     parser.add_argument('-b', '--batch', default=64, type=int,
-    					help='test batch size (default: 64)')
+                        help='test batch size (default: 64)')
     parser.add_argument('-t', '--test', default='CPN256x192', type=str,
-    					help='using which checkpoint to be tested (default: CPN256x192')
+                        help='using which checkpoint to be tested (default: CPN256x192')
     parser.add_argument('-r', '--result', default='result', type=str,
-    					help='path to save save result (default: result)')
+                        help='path to save save result (default: result)')
     main(parser.parse_args())
